requirements.txt
google-adk
google-generativeai
requests
# add any other packages you plan to use (e.g. for real API calls)
